requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
dotenv==0.0.5
flake8==7.1.1
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.1
